Add proper message size check for UDP (#92)
Added a path MTU config parameter for InitExportingProcess function.
For UDP transport, based on path MTU parameter, we make a decision on
the message size. This is optional parameter for TCP transport; TCP
supports max socket buffer size (65535).

+++ pkg/entities/testing/mock_set.go
@@ -63,10 +63,10 @@ func (mr *MockSetMockRecorder) AddRecord(arg0, arg1 interface{}) *gomock.Call {
 }
 
 // GetBuffLen mocks base method
-func (m *MockSet) GetBuffLen() uint16 {
+func (m *MockSet) GetBuffLen() int {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetBuffLen")
-	ret0, _ := ret[0].(uint16)
+	ret0, _ := ret[0].(int)
 	return ret0
 }
